requirements.txt
pandas>=1.3.0
numpy>=1.21.0
matplotlib>=3.4.0
seaborn>=0.11.0
mysql-connector-python>=8.0.26
scikit-learn>=0.24.0
jupyter>=1.0.0
notebook>=6.4.0
